requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-openai
langchain-pinecone
langchain-community
pypdf
python-dotenv
tiktoken